% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_badly.R
\name{parse_badly}
\alias{parse_badly}
\title{Parses --tblout format output from hammer}
\usage{
parse_badly(fn, trim.end = TRUE)
}
\arguments{
\item{fn}{a length-one character vector for the file path of the file to be parsed}
}
\description{
Parses --tblout format output from hammer
}
