% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_suffix_lines.R
\name{remove_suffix_lines}
\alias{remove_suffix_lines}
\title{Removes suffix lines from}
\usage{
remove_suffix_lines(d)
}
\arguments{
\item{d}{data frame produced by parse_badly()}
}
\description{
Removes suffix lines from
}
